fix: token refresh 로직 수정 및 적용
